Refactor the cluster list item components to be more flexible
