+++ src/components/MAPI/clusters/ClusterList/ClusterListItem.tsx
@@ -1,7 +1,10 @@
 import { useAuthProvider } from 'Auth/MAPI/MapiAuthProvider';
 import { push } from 'connected-react-router';
-import { Box } from 'grommet';
+import differenceInHours from 'date-fns/fp/differenceInHours';
+import toDate from 'date-fns-tz/toDate';
+import { Box, Card, CardBody, Text } from 'grommet';
 import ErrorReporter from 'lib/errors/ErrorReporter';
+import { relativeDate } from 'lib/helpers';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
 import RoutePath from 'lib/routePath';
 import { extractErrorMessage } from 'MAPI/organizations/utils';
@@ -15,12 +18,18 @@ import {
 import * as capiv1alpha3 from 'model/services/mapi/capiv1alpha3';
 import * as releasev1alpha1 from 'model/services/mapi/releasev1alpha1';
 import PropTypes from 'prop-types';
-import React, { useCallback, useEffect, useMemo, useRef } from 'react';
+import React, { useEffect, useMemo, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import { Link } from 'react-router-dom';
 import { OrganizationsRoutes } from 'shared/constants/routes';
 import { getProvider, getUserIsAdmin } from 'stores/main/selectors';
+import styled from 'styled-components';
 import useSWR from 'swr';
-import UIClusterListItem from 'UI/Display/MAPI/clusters/ClusterList/ClusterListItem';
+import Button from 'UI/Controls/Button';
+import ClusterIDLabel from 'UI/Display/Cluster/ClusterIDLabel';
+import ClusterListItemMainInfo from 'UI/Display/MAPI/clusters/ClusterList/ClusterListItemMainInfo';
+import ClusterListItemNodeInfo from 'UI/Display/MAPI/clusters/ClusterList/ClusterListItemNodeInfo';
+import ClusterListItemOptionalValue from 'UI/Display/MAPI/clusters/ClusterList/ClusterListItemOptionalValue';
 
 import {
   getWorkerNodesCount,
@@ -28,6 +37,33 @@ import {
   getWorkerNodesMemory,
 } from '../utils';
 import ClusterListItemStatus from './ClusterListItemStatus';
+
+const StyledLink = styled(Link)`
+  transition: box-shadow 0.1s ease-in-out;
+  display: block;
+  border-radius: ${(props) => props.theme.rounding}px;
+
+  :hover,
+  :focus {
+    text-decoration: none;
+    outline: none;
+    box-shadow: ${(props) =>
+      `0 0 0 1px ${props.theme.global.colors.text.dark}`};
+  }
+
+  &[aria-disabled='true'] {
+    cursor: default;
+
+    :hover,
+    :focus {
+      box-shadow: none;
+    }
+  }
+`;
+
+const GetStartedButton = styled(Button)`
+  text-transform: uppercase;
+`;
 
 interface IClusterListItemProps
   extends React.ComponentPropsWithoutRef<typeof Box> {
@@ -50,6 +86,9 @@ const ClusterListItem: React.FC<IClusterListItemProps> = ({
   const organization = cluster
     ? capiv1alpha3.getClusterOrganization(cluster)
     : undefined;
+
+  const creationDate = cluster?.metadata.creationTimestamp;
+  const deletionDate = cluster?.metadata.deletionTimestamp;
 
   const clusterPath = useMemo(() => {
     if (!organization || !name) return '';
@@ -123,9 +162,31 @@ const ClusterListItem: React.FC<IClusterListItemProps> = ({
         machineTypes.current
       );
 
+  const isAdmin = useSelector(getUserIsAdmin);
+  const provider = useSelector(getProvider);
+
+  const workerNodesCount = getWorkerNodesCount(nodePoolList?.items);
+
+  const isDeleting = Boolean(deletionDate);
+  const hasError = typeof nodePoolListError !== 'undefined';
+  const isLoading = typeof cluster === 'undefined';
+
+  const shouldDisplayGetStarted = useMemo(() => {
+    if (isDeleting || isLoading || !creationDate) return false;
+
+    const createDate = toDate(creationDate, { timeZone: 'UTC' });
+    const age = differenceInHours(createDate)(new Date());
+
+    // Cluster is older than 30 days.
+    // eslint-disable-next-line no-magic-numbers
+    return age < 30 * 24;
+  }, [creationDate, isDeleting, isLoading]);
+
   const dispatch = useDispatch();
 
-  const handleGetStartedClick = useCallback(() => {
+  const handleGetStartedClick = (e: React.MouseEvent<HTMLButtonElement>) => {
+    e.preventDefault();
+
     if (!organization || !name) return;
 
     const path = RoutePath.createUsablePath(
@@ -137,39 +198,105 @@ const ClusterListItem: React.FC<IClusterListItemProps> = ({
     );
 
     dispatch(push(path));
-  }, [dispatch, name, organization]);
-
-  const isAdmin = useSelector(getUserIsAdmin);
-  const provider = useSelector(getProvider);
+  };
 
   return (
-    <UIClusterListItem
-      href={clusterPath}
-      name={name}
-      namespace={cluster?.metadata.namespace}
-      description={description}
-      creationDate={cluster?.metadata.creationTimestamp}
-      deletionDate={cluster?.metadata.deletionTimestamp ?? null}
-      releaseVersion={releaseVersion}
-      k8sVersion={k8sVersion}
-      workerNodePoolsCount={nodePoolList?.items.length}
-      workerNodesCount={getWorkerNodesCount(nodePoolList?.items)}
-      workerNodesCPU={workerNodesCPU}
-      workerNodesMemory={workerNodesMemory}
-      workerNodesError={extractErrorMessage(nodePoolListError)}
-      onGetStartedClick={handleGetStartedClick}
-      additionalTitle={
-        cluster && (
-          <ClusterListItemStatus
-            cluster={cluster}
-            provider={provider}
-            isAdmin={isAdmin}
-            releases={releases}
-          />
-        )
-      }
-      {...props}
-    />
+    <StyledLink
+      to={isDeleting || isLoading ? '' : clusterPath}
+      aria-label={isLoading ? 'Loading cluster...' : `Cluster ${name}`}
+      aria-disabled={isDeleting || isLoading}
+      tabIndex={isDeleting || isLoading ? -1 : 0}
+    >
+      <Card
+        direction='row'
+        elevation='none'
+        overflow='visible'
+        background={isDeleting ? 'background-back' : 'background-front'}
+        round='xsmall'
+        pad='medium'
+        gap='small'
+        {...props}
+      >
+        <CardBody direction='row' gap='xsmall' wrap={true}>
+          <Box>
+            <ClusterListItemOptionalValue value={name}>
+              {(value) => (
+                <Text size='large' aria-label='Cluster name'>
+                  <ClusterIDLabel
+                    clusterID={value as string}
+                    copyEnabled={true}
+                  />
+                </Text>
+              )}
+            </ClusterListItemOptionalValue>
+          </Box>
+          <Box basis='80%'>
+            <Box direction='row' align='center' wrap={true} gap='small'>
+              <ClusterListItemOptionalValue value={description}>
+                {(value) => (
+                  <Text
+                    weight='bold'
+                    size='large'
+                    aria-label='Cluster description'
+                  >
+                    {value}
+                  </Text>
+                )}
+              </ClusterListItemOptionalValue>
+
+              {cluster && (
+                <ClusterListItemStatus
+                  cluster={cluster}
+                  provider={provider}
+                  isAdmin={isAdmin}
+                  releases={releases}
+                />
+              )}
+            </Box>
+
+            {isDeleting ? (
+              <Text color='text-xweak' aria-label='Cluster deletion date'>
+                Deleted {relativeDate(deletionDate)}
+              </Text>
+            ) : (
+              <ClusterListItemMainInfo
+                creationDate={creationDate}
+                releaseVersion={releaseVersion}
+                k8sVersion={k8sVersion}
+              />
+            )}
+
+            {!hasError && !isDeleting && (
+              <ClusterListItemNodeInfo
+                workerNodePoolsCount={nodePoolList?.items.length}
+                workerNodesCPU={workerNodesCPU}
+                workerNodesCount={workerNodesCount}
+                workerNodesMemory={workerNodesMemory}
+              />
+            )}
+
+            {hasError && !isDeleting && (
+              <Text color='status-critical' aria-label='Cluster load error'>
+                {extractErrorMessage(nodePoolListError)}
+              </Text>
+            )}
+          </Box>
+
+          {shouldDisplayGetStarted && (
+            <Box basis='10%'>
+              <GetStartedButton onClick={handleGetStartedClick}>
+                <i
+                  className='fa fa-start'
+                  aria-hidden={true}
+                  role='presentation'
+                />
+                Get Started
+              </GetStartedButton>
+            </Box>
+          )}
+        </CardBody>
+      </Card>
+    </StyledLink>
   );
 };
